// src/app/api-analytics/model/collectionValidationStats.ts
/**
 * MetaQS Analytics API
 *  * [**Real-Time API**](/metaqs-api/v1/docs) * [**LanguageTool API**](/metaqs-api/v1/languagetool/docs)     
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OehValidationError } from './oehValidationError';


export interface CollectionValidationStats { 
    title?: Array<OehValidationError>;
    keywords?: Array<OehValidationError>;
    description?: Array<OehValidationError>;
    edu_context?: Array<OehValidationError>;
}