// src/app/api-analytics/model/validationStatsResponseMaterialValidationStats.ts
/**
 * MetaQS Analytics API
 *  * [**Real-Time API**](/metaqs-api/v1/docs) * [**LanguageTool API**](/metaqs-api/v1/languagetool/docs)     
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MaterialValidationStats } from './materialValidationStats';


export interface ValidationStatsResponseMaterialValidationStats { 
    noderef_id: string;
    derived_at?: string;
    validation_stats: MaterialValidationStats;
}